+++ internal/cmd/plancal/dl.go
@@ -1,14 +1,12 @@
 package plancal
 
 import (
-	"fmt"
 	"os"
 	"regexp"
 	"strconv"
 	"strings"
 	"time"
 
-	"github.com/ag7if/go-files"
 	"github.com/spf13/cobra"
 
 	"github.com/derhabicht/planning-tools/internal/calendar"
@@ -16,7 +14,7 @@ import (
 )
 
 var dlCmd = &cobra.Command{
-	Use:   "dl [week]",
+	Use:   "dl [sprint|week]",
 	Short: "Create day labels",
 	Long:  ``,
 	Args:  cobra.RangeArgs(0, 1),
@@ -26,10 +24,31 @@ var dlCmd = &cobra.Command{
 func runDLCmd(cmd *cobra.Command, args []string) {
 	logger := logging.Logger{}
 
+	cards, err := cmd.Flags().GetBool("cards")
+	if err != nil {
+		logger.Debug().Err(err).Msg("failed to parse --cards flag")
+	}
+
+	labels, err := cmd.Flags().GetBool("labels")
+	if err != nil {
+		logger.Debug().Err(err).Msg("failed to parse --labels flag")
+	}
+
+	contexts, err := cmd.Flags().GetStringSlice("contexts")
+	if err != nil {
+		logger.Debug().Err(err).Msg("failed to parse --contexts flag")
+	}
+
+	if !(cards || labels) {
+		cards = true
+		labels = true
+	}
+
+	sprint := false
 	var year int
-	var week int
+	var period int
 	if len(args) > 0 {
-		weekre := `(\d{4})W(\d{2})`
+		weekre := `(\d{4})(S|W)(\d{2})`
 		re, err := regexp.Compile(weekre)
 		if err != nil {
 			panic(err)
@@ -39,31 +58,32 @@ func runDLCmd(cmd *cobra.Command, args []string) {
 
 		year, err = strconv.Atoi(res[1])
 		if err != nil {
-			logger.Error().Err(err).Msg("malformed week designator")
+			logger.Error().Err(err).Msg("malformed sprint/week designator")
 		}
 
-		week, err = strconv.Atoi(res[2])
+		if res[2] == "S" {
+			sprint = true
+		}
+
+		period, err = strconv.Atoi(res[3])
 		if err != nil {
-			logger.Error().Err(err).Msg("malformed week designator")
+			logger.Error().Err(err).Msg("malformed sprint/week designator")
 		}
 	} else {
 		now := time.Now()
-		year, week = now.ISOWeek()
+		year, period = now.ISOWeek()
 	}
 
-	outputFile, err := files.NewFile(fmt.Sprintf("DayLabels-%04dW%02d.pdf", year, week))
-	if err != nil {
-		logger.Error().Err(err).Msg("failed to create output file")
-		os.Exit(1)
-	}
-
-	err = calendar.BuildLabels(year, week, outputFile)
+	err = calendar.BuildDL(year, period, sprint, cards, labels, contexts)
 	if err != nil {
 		logger.Error().Err(err).Msg("failed to generate labels")
+		os.Exit(1)
 	}
-
 }
 
 func init() {
+	dlCmd.Flags().BoolP("cards", "c", false, "Generate cards")
+	dlCmd.Flags().BoolP("labels", "l", false, "Generate labels")
+	dlCmd.Flags().StringSliceP("contexts", "x", nil, "Contexts for generated cards")
 	rootCmd.AddCommand(dlCmd)
 }